refactor: put cursor restoration into the theme customizer
